+++ tutorial/react/getting-started/live-cursors/solved-diff/App.tsx
@@ -10,7 +10,7 @@ export default function App() {
   const publicApiKey = "{% LIVEBLOCKS_PUBLIC_KEY %}";
 
   return (
-    <LiveblocksProvider publicApiKey={publicApiKey}>
+    <LiveblocksProvider publicApiKey={publicApiKey} throttle={16}>
       <RoomProvider id={roomId} initialPresence={{ cursor: null }}>
         <ClientSideSuspense fallback={<div>Loading…</div>}>
           <Room />