+++ tutorial/react/getting-started/nesting-data-types/solved-diff/liveblocks.config.ts
@@ -1,11 +1,4 @@
-import { createClient, LiveList, LiveObject } from "@liveblocks/client";
-import { createRoomContext } from "@liveblocks/react";
-
-const client = createClient({
-  publicApiKey: "{% LIVEBLOCKS_PUBLIC_KEY %}",
-});
-
-type Presence = {};
+import { LiveList, LiveObject } from "@liveblocks/client";
 
 // Person type
 type Person = LiveObject<{
@@ -13,38 +6,13 @@ type Person = LiveObject<{
   age: number;
 }>;
 
-// Storage type
-type Storage = {
-  people: LiveList<Person>;
-};
+// Global types
+declare global {
+  interface Liveblocks {
+    Storage: {
+      people: LiveList<Person>;
+    };
+  }
+}
 
-type UserMeta = {};
-type RoomEvent = {};
-
-export const {
-  suspense: {
-    RoomProvider,
-    useRoom,
-    useMyPresence,
-    useUpdateMyPresence,
-    useSelf,
-    useOthers,
-    useOthersMapped,
-    useOthersConnectionIds,
-    useOther,
-    useBroadcastEvent,
-    useEventListener,
-    useErrorListener,
-    useStorage,
-    useObject,
-    useMap,
-    useList,
-    useBatch,
-    useHistory,
-    useUndo,
-    useRedo,
-    useCanUndo,
-    useCanRedo,
-    useMutation,
-  },
-} = createRoomContext<Presence, Storage, UserMeta, RoomEvent>(client);
+export {};