+++ packages/liveblocks-react/src/liveblocks.tsx
@@ -190,7 +190,7 @@ function makeExtrasForClient<U extends BaseUserMeta, M extends BaseMetadata>(
   const store = internals.cacheStore;
   const notifications = internals.notifications;
 
-  let fetchInboxNotificationsRequest: Promise<{
+  let fetchInboxNotifications$: Promise<{
     inboxNotifications: InboxNotificationData[];
     threads: ThreadData<M>[];
     deletedThreads: ThreadDeleteInfo[];
@@ -224,8 +224,8 @@ function makeExtrasForClient<U extends BaseUserMeta, M extends BaseMetadata>(
   async function fetchInboxNotifications(
     { retryCount }: { retryCount: number } = { retryCount: 0 }
   ) {
-    if (fetchInboxNotificationsRequest !== null) {
-      return fetchInboxNotificationsRequest;
+    if (fetchInboxNotifications$ !== null) {
+      return fetchInboxNotifications$;
     }
 
     store.setQueryState(INBOX_NOTIFICATIONS_QUERY, {
@@ -233,9 +233,9 @@ function makeExtrasForClient<U extends BaseUserMeta, M extends BaseMetadata>(
     });
 
     try {
-      fetchInboxNotificationsRequest = notifications.getInboxNotifications();
+      fetchInboxNotifications$ = notifications.getInboxNotifications();
 
-      const result = await fetchInboxNotificationsRequest;
+      const result = await fetchInboxNotifications$;
 
       store.updateThreadsAndNotifications(
         result.threads,
@@ -247,7 +247,7 @@ function makeExtrasForClient<U extends BaseUserMeta, M extends BaseMetadata>(
 
       /**
        * We set the `lastRequestedAt` to the timestamp returned by the current request if:
-       * 1. The `lastRequestedAt`has not been set
+       * 1. The `lastRequestedAt` has not been set
        * OR
        * 2. The current `lastRequestedAt` is older than the timestamp returned by the current request
        */
@@ -260,7 +260,7 @@ function makeExtrasForClient<U extends BaseUserMeta, M extends BaseMetadata>(
 
       poller.start(POLLING_INTERVAL);
     } catch (er) {
-      fetchInboxNotificationsRequest = null;
+      fetchInboxNotifications$ = null;
 
       // Retry the action using the exponential backoff algorithm
       retryError(() => {