+++ packages/liveblocks-react/src/room.tsx
@@ -76,6 +76,7 @@ import { selectedThreads } from "./comments/lib/selected-threads";
 import { retryError } from "./lib/retry-error";
 import { useInitial } from "./lib/use-initial";
 import { useLatest } from "./lib/use-latest";
+import { use } from "./lib/use-polyfill";
 import {
   createSharedContext,
   LiveblocksProviderWithClient,
@@ -2286,23 +2287,12 @@ function ensureNotServerSide(): void {
   }
 }
 
-function useSuspendUntilPresenceLoaded(): void {
-  const room = useRoom();
-  if (room.getSelf() !== null) {
-    return;
-  }
-
+function useSuspendUntilPresenceReady(): void {
+  // Throw an error if we're calling this on the server side
   ensureNotServerSide();
 
-  // Throw a _promise_. Suspense will suspend the component tree until either
-  // until either a presence update event, or a connection status change has
-  // happened. After that, it will render this component tree again and
-  // re-evaluate the .getSelf() condition above, or re-suspend again until
-  // such event happens.
-  throw new Promise<void>((res) => {
-    room.events.self.subscribeOnce(() => res());
-    room.events.status.subscribeOnce(() => res());
-  });
+  const room = useRoom();
+  use(room.waitUntilPresenceReady());
 }
 
 function useSelfSuspense<P extends JsonObject, U extends BaseUserMeta>(): User<
@@ -2317,7 +2307,7 @@ function useSelfSuspense<P extends JsonObject, U extends BaseUserMeta, T>(
   selector?: (me: User<P, U>) => T,
   isEqual?: (prev: T, curr: T) => boolean
 ): T | User<P, U> {
-  useSuspendUntilPresenceLoaded();
+  useSuspendUntilPresenceReady();
   return useSelf(
     selector as (me: User<P, U>) => T,
     isEqual as (prev: T | null, curr: T | null) => boolean
@@ -2336,7 +2326,7 @@ function useOthersSuspense<P extends JsonObject, U extends BaseUserMeta, T>(
   selector?: (others: readonly User<P, U>[]) => T,
   isEqual?: (prev: T, curr: T) => boolean
 ): T | readonly User<P, U>[] {
-  useSuspendUntilPresenceLoaded();
+  useSuspendUntilPresenceReady();
   return useOthers(
     selector as (others: readonly User<P, U>[]) => T,
     isEqual as (prev: T, curr: T) => boolean
@@ -2358,7 +2348,7 @@ function useOthersSuspense<P extends JsonObject, U extends BaseUserMeta, T>(
  * // [2, 4, 7]
  */
 function useOthersConnectionIdsSuspense(): readonly number[] {
-  useSuspendUntilPresenceLoaded();
+  useSuspendUntilPresenceReady();
   return useOthersConnectionIds();
 }
 
@@ -2370,7 +2360,7 @@ function useOthersMappedSuspense<
   itemSelector: (other: User<P, U>) => T,
   itemIsEqual?: (prev: T, curr: T) => boolean
 ): ReadonlyArray<readonly [connectionId: number, data: T]> {
-  useSuspendUntilPresenceLoaded();
+  useSuspendUntilPresenceReady();
   return useOthersMapped(itemSelector, itemIsEqual);
 }
 
@@ -2379,31 +2369,23 @@ function useOtherSuspense<P extends JsonObject, U extends BaseUserMeta, T>(
   selector: (other: User<P, U>) => T,
   isEqual?: (prev: T, curr: T) => boolean
 ): T {
-  useSuspendUntilPresenceLoaded();
+  useSuspendUntilPresenceReady();
   return useOther(connectionId, selector, isEqual);
 }
 
-function useSuspendUntilStorageLoaded(): void {
-  const room = useRoom();
-  if (room.getStorageSnapshot() !== null) {
-    return;
-  }
-
+function useSuspendUntilStorageReady(): void {
+  // Throw an error if we're calling this on the server side
   ensureNotServerSide();
 
-  // Throw a _promise_. Suspense will suspend the component tree until this
-  // promise resolves (aka until storage has loaded). After that, it will
-  // render this component tree again.
-  throw new Promise<void>((res) => {
-    room.events.storageDidLoad.subscribeOnce(() => res());
-  });
+  const room = useRoom();
+  use(room.waitUntilStorageReady());
 }
 
 function useStorageSuspense<S extends LsonObject, T>(
   selector: (root: ToImmutable<S>) => T,
   isEqual?: (prev: T, curr: T) => boolean
 ): T {
-  useSuspendUntilStorageLoaded();
+  useSuspendUntilStorageReady();
   return useStorage(
     selector,
     isEqual as (prev: T | null, curr: T | null) => boolean
@@ -2416,7 +2398,7 @@ function useStorageSuspense<S extends LsonObject, T>(
  * indicator.
  */
 function useStorageStatusSuspense(): StorageStatusSuccess {
-  useSuspendUntilStorageLoaded();
+  useSuspendUntilStorageReady();
   const room = useRoom();
   const subscribe = room.events.storageStatus.subscribe;
   const getSnapshot = room.getStorageStatus as () => StorageStatusSuccess;